requirements.txt
# Ficheiro: requirements.txt (VERSÃO DEFINITIVA E ALINHADA)

# --- Core Libraries ---
numpy
pandas
setuptools

# --- Machine Learning & AI ---
optuna
# scikit-learn
# optuna-integration # Comentado - Descomente apenas se o seu código o usar explicitamente
# stable-baselines3 # Comentado - Para uso futuro em Reinforcement Learning
# gymnasium # Comentado - Para uso futuro em Reinforcement Learning

# --- Data & APIs ---
python-binance
requests # Essencial para o pipeline de sentimento
sqlalchemy
psycopg2-binary
alembic
httpx
fastapi
uvicorn

# --- Technical Analysis ---
# ta # Comentado - pandas-ta geralmente cobre tudo
pandas_ta # Sem versão para instalar a mais recente e compatível

# --- Config & Utilities ---
typer
python-dotenv
tqdm
joblib
tabulate
questionary

# --- Plotting ---
matplotlib
seaborn
textual
textual-plotext
textual-timepiece
whenever
pytz

# --- Testing ---
pytest
aiocsv
psutil
